configuration of astro.config, remove elements that don't need it in the project, creation and rename of elements that need to change, modification of the update to use customElements, creation of json resource to get data
