+++ src/assets/json/NorthAmerica.json
@@ -1,0 +1,4 @@
+{
+    "Carolina":{"name":"hola"},
+    "carolina2":{}
+}
