+++ src/layouts/MainLayout.astro
@@ -13,7 +13,7 @@ const {title} = Astro.props
 <html lang="en">
 <head>
     <ClientRouter fallback='animate'/>
-    <link href="/node_modules/material-icons/iconfont/material-icons.css" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css2?family=Material+Icons" rel="stylesheet">
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta name="description" content="Main page of 100 destination" />
