+++ src/types/continents.d.ts
@@ -1,0 +1,12 @@
+type Dictionary<Key extends keyof any, Value> = {
+    [key in Key]: Value
+}
+interface Coord {
+    x:number
+    y:number
+}
+interface Continent {
+    coordinate: Coord,
+    abbr: string,
+    continentImg:string
+}
